# Conflicts:
#	CookieEngine/CMakeLists.txt
#	CookieEngine/include/Resources/Scene.hpp
#	CookieEngine/src/Editor.cpp
#	CookieEngine/src/Game.cpp
#	CookieEngine/src/UI/UIeditor.cpp
#	CookieEngine/src/UI/Widgets/EditorWs/TextureEditorWidget.cpp
#	CookieEngine/src/UI/Widgets/GameportWidget.cpp
#	CookieEngine/src/main.cpp

+++ CookieEngine/deps/include/reactphysics3d/collision/OverlapCallback.h
@@ -186,7 +186,7 @@ class OverlapCallback {
 
 // Return the number of overlapping pairs of bodies
 inline uint OverlapCallback::CallbackData::getNbOverlappingPairs() const {
-    return mContactPairsIndices.size() + mLostContactPairsIndices.size();
+    return (uint)(mContactPairsIndices.size() + mLostContactPairsIndices.size());
 }
 
 // Return a given overlapping pair of bodies
